Se arreglaron las hitboxes del robot para que sea mas jugable. Tambien se maneja diferente la dificultad del juego. Es mas entendible y ya se implemento el cambio de dificultad en el personaje Robot.

+++ app/src/main/java/com/jgame/elements/AnimationData.java
@@ -38,7 +38,7 @@ public class AnimationData {
         this.loops = loops;
     }
 
-    public void updateFrameData(byte framesPerSprite){
+    public void updateFrameData(int framesPerSprite){
         this.framesPerSprite = framesPerSprite;
     }
